requirements.txt
openai
requests
readability-lxml
html2text
sentence-transformers
langchain
langchain-community